+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.1 [unreleased]
 
+- Added `correlation` method
 - Added support for API keys
 
 ## 0.1.0